Examples/Python/requirements.txt
numpy==1.24.3
Pillow==9.5.0
pytest==7.3.1
pytest-benchmark==4.0.0
python-dotenv==1.0.0
